Minion optimization
1 Move some functions to sub-module
